src/main/java/lpf/learn/leetcode/tags/dp/CountSquareSubmatricesWithAllOnes.java
package lpf.learn.leetcode.tags.dp;

import java.util.stream.IntStream;

/** [1277]统计全为 1 的正方形子矩阵
 * 给你一个 m * n 的矩阵，矩阵中的元素不是 0 就是 1，请你统计并返回其中完全由 1 组成的 正方形 子矩阵的个数。
 *
 * 示例 1：
 * 输入：matrix =
 * [
 *   [0,1,1,1],
 *   [1,1,1,1],
 *   [0,1,1,1]
 * ]
 * 输出：15
 * 解释：
 * 边长为 1 的正方形有 10 个。
 * 边长为 2 的正方形有 4 个。
 * 边长为 3 的正方形有 1 个。
 * 正方形的总数 = 10 + 4 + 1 = 15.
 *
 * 示例 2：
 * 输入：matrix =
 * [
 *  [1,0,1],
 *  [1,1,0],
 *  [1,1,0]
 * ]
 * 输出：7
 * 解释：
 * 边长为 1 的正方形有 6 个。
 * 边长为 2 的正方形有 1 个。
 * 正方形的总数 = 6 + 1 = 7.
 *
 * 提示：
 * 1 <= arr.length <= 300
 * 1 <= arr[0].length <= 300
 * 0 <= arr[i][j] <= 1
 */
public class CountSquareSubmatricesWithAllOnes {
    public int countSquares(int[][] matrix) {
        int m = matrix.length;
        int n = matrix[0].length;
        // dp[i][j] 表示 右下角以（i，j）为顶点的正方形的最大边长

        int result = IntStream.of(matrix[0]).sum();
        for (int i = 1; i < m; i++) {
            if (matrix[i][0] == 0) {
                continue;
            }
            result++;
        }

        for (int i = 1; i < m; i++) {
            for (int j = 1; j < n; j++) {
                if (matrix[i][j] == 0) {
                    continue;
                }
                int l = Math.min(matrix[i - 1][j], Math.min(matrix[i - 1][j - 1], matrix[i][j - 1])) + 1;

                matrix[i][j] = l;
                result += l;
            }
        }

        return result;
    }
}
